Implemented load game function.
Logged in users can now load saved games.
Also the save and load buttons are hidden when the user is not logged
in.
If an invalid mapId is entered in the url an error page will handle it
Displays logged in user name next to logout button is top navigation
bar. This was really so I could grab it when updating the score board.
Need to save score to user docuemnt still.